[PR Feedback] Rename "events" module to "event"
