+++ components/sup/src/lib.rs
@@ -94,7 +94,7 @@ pub mod command;
 pub mod config;
 pub mod ctl_gateway;
 pub mod error;
-mod events;
+mod event;
 pub mod http_gateway;
 pub mod manager;
 mod sys;
