+++ components/sup/src/manager/mod.rs
@@ -49,9 +49,9 @@ use crate::{census::{CensusRing,
             error::{Error,
                     Result,
                     SupError},
-            events::{self,
-                     EventConnectionInfo,
-                     EventCore},
+            event::{self,
+                    EventConnectionInfo,
+                    EventCore},
             feat,
             http_gateway,
             VERSION};
@@ -492,7 +492,7 @@ impl Manager {
             let ec = EventCore { supervisor_id: sys.member_id.clone(), };
             // TODO: Determine what the actual connection parameters
             // should be, and process them at some point before here.
-            events::init_stream(EventConnectionInfo::default(), ec);
+            event::init_stream(EventConnectionInfo::default(), ec);
         }
 
         Ok(Manager { state: Arc::new(ManagerState { cfg: cfg_static,
@@ -665,9 +665,9 @@ impl Manager {
             .expect("Updater lock poisoned")
             .add(&service);
 
-        events::publish(events::ServiceStarted { ident:         &service.pkg.ident,
-                                                 spec_ident:    &spec.ident,
-                                                 service_group: &service.service_group, });
+        event::publish(event::ServiceStarted { ident:         &service.pkg.ident,
+                                               spec_ident:    &spec.ident,
+                                               service_group: &service.service_group, });
 
         self.state
             .services
@@ -1187,7 +1187,7 @@ impl Manager {
         // TODO (CM): But only if we're not going down for a restart.
         let ident = service.spec_ident.clone();
         let stop_it = service.stop().then(move |_| {
-                                        events::publish(events::ServiceStopped {
+                                        event::publish(event::ServiceStopped {
                 ident: &service.pkg.ident,
                 //                spec_ident: &service.spec.ident,
                 service_group: &service.service_group,
